+++ src/main/java/com/haisheng/framework/testng/bigScreen/jiaochen/wm/sense/pc/receptionmanager/PurchaseTemporaryPackage.java
@@ -21,7 +21,7 @@ public class PurchaseTemporaryPackage extends BaseScene {
     private final String extendedInsuranceYear;
     private final String extendedInsuranceCopies;
     private final Integer type;
-    private final Long receptionId;
+    private final Integer receptionId;
     private final Long customerId;
 
     @Override
